# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Renewal Scout API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services/{id}/current-deal:
    get:
      operationId: getCurrentDeal
      tags: [deals]
      summary: Get the current deal for a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Current deal fields with provenance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentDeal"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCurrentDeal
      tags: [deals]
      summary: Save or update the current deal for a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CurrentDealInput"
      responses:
        "200":
          description: Updated current deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentDeal"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /services/{id}/extract-document:
    post:
      operationId: extractDocument
      tags: [deals]
      summary: >
        Upload a document (PDF, JPG, PNG ≤ 10 MB) and extract deal fields via AI.
        The document is sent to the OpenAI API; document bytes are never persisted.
        Returns a draft with all fields marked extracted_unconfirmed — user must
        confirm before values are used in research.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                document:
                  type: string
                  format: binary
              required:
                - document
      responses:
        "200":
          description: Extraction draft — all fields are extracted_unconfirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractionDraft"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: File too large (> 10 MB)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported file type (only PDF, JPG, PNG accepted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /services/{id}/extraction-draft/{extractionId}/confirm:
    put:
      operationId: confirmExtractionDraft
      tags: [deals]
      summary: >
        Confirm, correct, or delete extracted field values.
        Confirmed fields are saved to current_deals with source extracted_confirmed.
        Deleted fields are removed from the draft.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
        - name: extractionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractionConfirmInput"
      responses:
        "200":
          description: Updated current deal after confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentDeal"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Service or extraction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ProvenanceSource:
      type: string
      enum:
        - user
        - extracted_confirmed
        - extracted_unconfirmed
        - unknown
    ProvenanceField:
      type: object
      properties:
        value: {}
        source:
          $ref: "#/components/schemas/ProvenanceSource"
      required:
        - source
    CurrentDeal:
      type: object
      properties:
        serviceId:
          type: number
        fields:
          type: object
          description: Deal fields — each value includes provenance
          additionalProperties:
            $ref: "#/components/schemas/ProvenanceField"
        lastConfirmedAt:
          type: ["string", "null"]
        updatedAt:
          type: string
      required:
        - serviceId
        - fields
        - updatedAt
    CurrentDealInput:
      type: object
      properties:
        fields:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ProvenanceField"
      required:
        - fields

    # ─── Document extraction ────────────────────────────────────────────────────
    ExtractionDraft:
      type: object
      properties:
        extractionId:
          type: string
        serviceId:
          type: number
        fields:
          type: object
          description: >
            Extracted deal fields — all have source=extracted_unconfirmed
            until the user confirms them.
          additionalProperties:
            $ref: "#/components/schemas/ProvenanceField"
        extractedAt:
          type: string
        aiDisclosure:
          type: string
          description: Human-readable notice that the document was sent to the OpenAI API
      required:
        - extractionId
        - serviceId
        - fields
        - extractedAt
        - aiDisclosure
    ExtractionConfirmInput:
      type: object
      properties:
        confirmedFields:
          type: object
          description: >
            Fields to save with source=extracted_confirmed.
            Pass corrected values here — the user's edits override AI values.
          additionalProperties:
            $ref: "#/components/schemas/ProvenanceField"
        deletedFields:
          type: array
          items:
            type: string
          description: Field names to remove from current_deals entirely
      required:
        - confirmedFields

    # ─── Completeness ───────────────────────────────────────────────────────────
...