# lib/api-zod/src/generated/types/extractionDraft.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Renewal Scout API
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractionDraftFields } from './extractionDraftFields';

export interface ExtractionDraft {
  extractionId: string;
  serviceId: number;
  /** Extracted deal fields — all have source=extracted_unconfirmed until the user confirms them. */
  fields: ExtractionDraftFields;
  extractedAt: string;
  /** Human-readable notice that the document was sent to the OpenAI API */
  aiDisclosure: string;
}