# lib/api-zod/src/generated/types/provenanceSource.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Renewal Scout API
 * OpenAPI spec version: 0.1.0
 */

export type ProvenanceSource = typeof ProvenanceSource[keyof typeof ProvenanceSource];


export const ProvenanceSource = {
  user: 'user',
  extracted_confirmed: 'extracted_confirmed',
  extracted_unconfirmed: 'extracted_unconfirmed',
  unknown: 'unknown',
} as const;